pmt_core_pkg/pyproject.toml
[project]
name = "pmt-core"
version = "0.1.0"
description = "Shared business logic for Portfolio Management Tool"
requires-python = ">=3.11"
dependencies = [
    "pandas",
    "pyodbc",
    "python-dotenv",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["pmt_core*"]
